+++ packages/editor/adf-schema/src/__tests__/unit/schema/nodes/status.ts
@@ -1,6 +1,7 @@
 import { name } from '../../../../../package.json';
 import { schema, toDOM, fromHTML } from '../../../../../test-helpers';
 import { status } from '../../../../../src';
+import { StatusLocalIdRegex } from '@atlaskit/editor-test-helpers';
 
 describe(`${name}/schema status node`, () => {
   describe('parse html', () => {
@@ -28,10 +29,14 @@ describe(`${name}/schema status node`, () => {
         schema,
       );
       const node = doc.firstChild!.firstChild!;
-      expect(node.attrs.color).toEqual(color);
-      expect(node.attrs.localId).toEqual(localId);
-      expect(node.attrs.text).toEqual('In progress');
-      expect(node.attrs.style).toEqual(style);
+      expect(node.attrs).toMatchObject({
+        text: 'In progress',
+        color,
+        localId: expect.stringMatching(StatusLocalIdRegex),
+        style,
+      });
+
+      expect(node.attrs.localId).not.toEqual(localId);
     });
 
     it('gets attributes from html without optional ones', () => {
@@ -48,12 +53,13 @@ describe(`${name}/schema status node`, () => {
         schema,
       );
       const node = doc.firstChild!.firstChild!;
-      expect(node.attrs.color).toEqual(color);
-      expect(node.attrs.text).toEqual('In progress');
-      expect(node.attrs.style).toEqual(null);
-      expect(node.attrs.localId).toHaveLength(
-        '7f4189c0-89f2-4f0e-a439-3fa9e57934fa'.length,
-      ); // random UUID, just match the length
+
+      expect(node.attrs).toMatchObject({
+        text: 'In progress',
+        color,
+        localId: expect.stringMatching(StatusLocalIdRegex),
+        style: null,
+      });
     });
   });
 
@@ -85,7 +91,15 @@ describe(`${name}/schema status node`, () => {
       const dom = toDOM(node, schema).firstChild as HTMLElement;
       const parsedNode = fromHTML(dom.outerHTML, schema).firstChild!
         .firstChild!;
-      expect(parsedNode).toEqual(node);
+
+      expect(parsedNode.attrs).toMatchObject({
+        text: 'In progress',
+        color: 'blue',
+        localId: expect.stringMatching(StatusLocalIdRegex),
+        style: 'bold',
+      });
+
+      expect(parsedNode.attrs.localId).not.toEqual(attrs.localId);
     });
 
     it('converts html status attributes to node attributes without style', () => {