+++ packages/editor/editor-core/src/__tests__/integration/card/inline-2.ts
@@ -22,40 +22,38 @@ const moveUp = (page, selector) => {
   return page.browser.addValue(selector, keys);
 };
 
-[fullpage].forEach(editor => {
-  BrowserTestCase(
-    `inline-2.ts: pasting an link then typing still converts to inline card`,
-    {
-      skip: ['ie', 'safari'],
-    },
-    async client => {
-      let browser = new Page(client);
+BrowserTestCase(
+  `inline-2.ts: pasting an link then typing still converts to inline card`,
+  {
+    skip: ['ie', 'safari'],
+  },
+  async client => {
+    let browser = new Page(client);
 
-      // copy stuff to clipboard
-      await browser.goto(clipboardHelper);
-      await browser.isVisible(clipboardInput);
-      await browser.type(clipboardInput, 'https://www.atlassian.com');
-      await browser.click(copyAsPlaintextButton);
+    // copy stuff to clipboard
+    await browser.goto(clipboardHelper);
+    await browser.isVisible(clipboardInput);
+    await browser.type(clipboardInput, 'https://www.atlassian.com');
+    await browser.click(copyAsPlaintextButton);
 
-      // open up editor
-      await browser.goto(editor.path);
-      await browser.waitForSelector(editor.placeholder);
-      await browser.click(editor.placeholder);
-      await browser.waitForSelector(editable);
+    // open up editor
+    await browser.goto(fullpage.path);
+    await browser.waitForSelector(fullpage.placeholder);
+    await browser.click(fullpage.placeholder);
+    await browser.waitForSelector(editable);
 
-      // type some text into the paragraph first
-      await browser.type(editable, 'hello have a link ');
+    // type some text into the paragraph first
+    await browser.type(editable, 'hello have a link ');
 
-      // paste the link
-      await browser.paste(editable);
+    // paste the link
+    await browser.paste(editable);
 
-      // type some text around it
-      await browser.type(editable, 'more typing');
-      await moveUp(browser, editable);
-      await browser.type(editable, 'more typing ');
+    // type some text around it
+    await browser.type(editable, 'more typing');
+    await moveUp(browser, editable);
+    await browser.type(editable, 'more typing ');
 
-      const doc = await browser.$eval(editable, getDocFromElement);
-      expect(doc).toMatchDocSnapshot();
-    },
-  );
-});
+    const doc = await browser.$eval(editable, getDocFromElement);
+    expect(doc).toMatchDocSnapshot();
+  },
+);