+++ packages/editor/editor-core/src/__tests__/unit/plugins/table/ui/ColumnControls.tsx
@@ -1,11 +1,5 @@
 import * as React from 'react';
-import {
-  selectTable,
-  getCellsInColumn,
-  getSelectionRect,
-} from 'prosemirror-utils';
-import { Node } from 'prosemirror-model';
-import { CellSelection } from 'prosemirror-tables';
+import { selectTable, getSelectionRect } from 'prosemirror-utils';
 import {
   doc,
   p,
@@ -17,6 +11,7 @@ import {
   td,
   thEmpty,
   mountWithIntl,
+  selectColumns,
 } from '@atlaskit/editor-test-helpers';
 
 import { pluginKey } from '../../../../../plugins/table/pm-plugins/main';
@@ -32,22 +27,6 @@ const ColumnControlsButtonWrap = `.${ClassName.COLUMN_CONTROLS_BUTTON_WRAP}`;
 const DeleteColumnButton = `.${ClassName.CONTROLS_DELETE_BUTTON_WRAP}`;
 const InsertColumnButton = `.${ClassName.CONTROLS_INSERT_BUTTON_WRAP}`;
 const InsertColumnButtonInner = `.${ClassName.CONTROLS_INSERT_BUTTON_INNER}`;
-
-const selectColumns = columnIdxs => tr => {
-  const cells: { pos: number; start: number; node: Node }[] = columnIdxs.reduce(
-    (acc, colIdx) => {
-      const colCells = getCellsInColumn(colIdx)(tr.selection);
-      return colCells ? acc.concat(colCells) : acc;
-    },
-    [],
-  );
-
-  if (cells) {
-    const $anchor = tr.doc.resolve(cells[0].pos);
-    const $head = tr.doc.resolve(cells[cells.length - 1].pos);
-    return tr.setSelection(new CellSelection($anchor, $head));
-  }
-};
 
 describe('ColumnControls', () => {
   const editor = (doc: any) =>
@@ -76,7 +55,6 @@ describe('ColumnControls', () => {
           column,
         );
         floatingControls.unmount();
-        editorView.destroy();
       });
     });
   });
@@ -269,7 +247,7 @@ describe('ColumnControls', () => {
 
       expect(floatingControls.find(InsertColumnButton).length).toBe(3);
 
-      editorView.dispatch(selectColumns([0, 1])(editorView.state.tr));
+      selectColumns([0, 1])(editorView.state, editorView.dispatch);
 
       // set numberOfColumns prop to trick shouldComponentUpdate and force re-render
       floatingControls.setProps({ numberOfColumns: 3 });
@@ -303,7 +281,7 @@ describe('ColumnControls', () => {
         />,
       );
 
-      editorView.dispatch(selectColumns([0, 1])(editorView.state.tr));
+      selectColumns([0, 1])(editorView.state, editorView.dispatch);
 
       // set numberOfColumns prop to trick shouldComponentUpdate and force re-render
       floatingControls.setProps({ numberOfColumns: 3 });
@@ -384,7 +362,7 @@ describe('ColumnControls', () => {
         ),
       );
 
-      editorView.dispatch(selectColumns([0])(editorView.state.tr));
+      selectColumns([0])(editorView.state, editorView.dispatch);
       const target = document.querySelectorAll(
         `.${ClassName.COLUMN_CONTROLS} .${ClassName.CONTROLS_BUTTON}`,
       )[2];
@@ -405,7 +383,7 @@ describe('ColumnControls', () => {
         ),
       );
 
-      editorView.dispatch(selectColumns([2])(editorView.state.tr));
+      selectColumns([2])(editorView.state, editorView.dispatch);
       const target = document.querySelectorAll(
         `.${ClassName.COLUMN_CONTROLS} .${ClassName.CONTROLS_BUTTON}`,
       )[0];